# back/java/src/main/java/com/resumethinking/platform/resumes/ResumeController.java
package com.resumethinking.platform.resumes;
import com.fasterxml.jackson.annotation.JsonIgnoreProperties; import com.resumethinking.platform.auth.UserRole; import com.resumethinking.platform.crypto.AesGcmCryptoService; import jakarta.validation.Valid; import jakarta.validation.constraints.*; import org.springframework.beans.factory.annotation.Autowired; import org.springframework.data.domain.PageRequest; import org.springframework.data.domain.Sort; import org.springframework.http.*; import org.springframework.web.bind.annotation.*; import org.springframework.web.multipart.MultipartFile; import java.time.Instant; import java.util.List;
@RestController @RequestMapping("/api/v2") public class ResumeController {
 private final ResumeLifecycleService service; private final AesGcmCryptoService crypto; public ResumeController(ResumeLifecycleService service){this(service,null);} @Autowired public ResumeController(ResumeLifecycleService service,AesGcmCryptoService crypto){this.service=service;this.crypto=crypto;}
 @PostMapping(value="/resumes",consumes=MediaType.MULTIPART_FORM_DATA_VALUE) public ResponseEntity<Response> upload(@RequestAttribute("actorId") String actor,@RequestAttribute("role") UserRole role,@RequestPart("file") MultipartFile file,@RequestParam(required=false) String title)throws java.io.IOException{String name=file.getOriginalFilename()==null?"resume.txt":file.getOriginalFilename();String lower=name.toLowerCase(java.util.Locale.ROOT);Resume.SourceType type;if(lower.endsWith(".txt"))type=Resume.SourceType.TXT;else if(lower.endsWith(".docx"))type=Resume.SourceType.DOCX;else throw new IllegalArgumentException("UNSUPPORTED_FILE");if(file.getSize()>5_000_000)throw new IllegalArgumentException("PAYLOAD_TOO_LARGE");if(crypto==null)throw new IllegalStateException("encryption unavailable");String safeTitle=(title==null||title.isBlank())?name.substring(0,Math.min(200,name.length())):title;if(safeTitle.length()>200)throw new IllegalArgumentException("VALIDATION_ERROR");var encrypted=crypto.encryptBytes(file.getBytes());return ResponseEntity.status(HttpStatus.CREATED).body(Response.from(service.upload(actor,role,safeTitle,type,encrypted.ciphertext(),encrypted.nonce())));}
 @GetMapping("/resumes") public Page list(@RequestAttribute("actorId") String actorId,@RequestAttribute("role") UserRole role,@RequestParam(defaultValue="1")int page,@RequestParam(defaultValue="20")int pageSize){validatePage(page,pageSize);return Page.from(service.listActive(actorId,role,PageRequest.of(page-1,pageSize,Sort.by("id"))),page,pageSize);}
 @GetMapping("/resumes/{id}") public Response get(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@PathVariable String id){return Response.from(service.getActive(id,actorId,role));}
 @DeleteMapping("/resumes/{id}") public Response delete(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@PathVariable String id,@Valid@RequestBody DeleteRequest body){return Response.from(service.softDelete(new DeleteResumeCommand(id,actorId,role,body.confirmationText(),body.expectedVersion())));}
 @PostMapping("/recovery/resumes/{id}/restore") public Response restore(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@PathVariable String id,@Valid@RequestBody RestoreRequest body){if(role!=UserRole.USER)throw new org.springframework.security.access.AccessDeniedException("FORBIDDEN");return Response.from(service.recover(id,actorId,role,body.expectedVersion()));}
 @PostMapping("/admin/recovery/resumes/{id}/restore") public Response adminRestore(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@PathVariable String id,@Valid@RequestBody RestoreRequest body){if(role!=UserRole.ADMIN)throw new org.springframework.security.access.AccessDeniedException("FORBIDDEN");return Response.from(service.recover(id,actorId,role,body.expectedVersion()));}
 @DeleteMapping("/admin/recovery/resumes/{id}") public Response adminDelete(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@PathVariable String id,@Valid@RequestBody DeleteRequest body){if(role!=UserRole.ADMIN)throw new org.springframework.security.access.AccessDeniedException("FORBIDDEN");return Response.from(service.softDelete(new DeleteResumeCommand(id,actorId,role,body.confirmationText(),body.expectedVersion())));}
 @GetMapping("/recovery/resumes") public Page recovery(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@RequestParam(defaultValue="1")int page,@RequestParam(defaultValue="20")int pageSize){if(role!=UserRole.USER)throw new org.springframework.security.access.AccessDeniedException("FORBIDDEN");validatePage(page,pageSize);return Page.from(service.listRecoverable(actorId,role,null,PageRequest.of(page-1,pageSize,Sort.by("id"))),page,pageSize);}
 @GetMapping("/admin/recovery/resumes") public Page adminRecovery(@RequestAttribute("actorId")String actorId,@RequestAttribute("role")UserRole role,@RequestParam(required=false)String ownerId,@RequestParam(defaultValue="1")int page,@RequestParam(defaultValue="20")int pageSize){if(role!=UserRole.ADMIN)throw new org.springframework.security.access.AccessDeniedException("FORBIDDEN");validatePage(page,pageSize);return Page.from(service.listRecoverable(actorId,role,ownerId,PageRequest.of(page-1,pageSize,Sort.by("id"))),page,pageSize);}
 @JsonIgnoreProperties(ignoreUnknown=false) public record DeleteRequest(@NotBlank String confirmationText,@NotNull @Min(0) Long expectedVersion){} @JsonIgnoreProperties(ignoreUnknown=false) public record RestoreRequest(@NotNull @Min(0) Long expectedVersion){} private static void validatePage(int page,int size){if(page<1||size<1||size>100)throw new IllegalArgumentException("VALIDATION_ERROR");}
 public record Page(List<Response> items,int page,int pageSize,long totalItems,int totalPages){static Page from(org.springframework.data.domain.Page<Resume> data,int page,int size){return new Page(data.getContent().stream().map(Response::from).toList(),page,size,data.getTotalElements(),data.getTotalPages());}}
 public record Response(String id,String ownerId,String title,Resume.SourceType sourceType,int status,VisibilityState visibilityState,long version,Instant visibleUntil,Instant softDeletedAt,Instant archivedAt,Instant restoredAt,Instant createdAt,Instant updatedAt){static Response from(Resume r){return new Response(r.getId(),r.getOwnerId(),r.getTitle(),r.getSourceType(),r.getStatus(),r.getVisibilityState(),r.getVersion(),r.getVisibleUntil(),r.getSoftDeletedAt(),r.getArchivedAt(),r.getRestoredAt(),r.getCreatedAt(),r.getUpdatedAt());}static Response from(ResumeView r){return new Response(r.id(),r.ownerId(),r.title(),r.sourceType(),r.status(),r.visibilityState(),r.version(),r.visibleUntil(),r.softDeletedAt(),r.archivedAt(),r.restoredAt(),r.createdAt(),r.updatedAt());}}
}