# contracts/openapi/v2/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Matching Platform Java API
  version: 2.1.4
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  description: |
    Versioned public API owned by the Java service. Resume and match data is
    scoped to the authenticated actor and is never returned in error messages.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/resumes:
    get:
      tags: [Resumes]
      summary: List active resumes
      operationId: listResumes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Active resumes visible to the authenticated actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    post:
      tags: [Resumes]
      summary: Upload a TXT or DOCX resume
      operationId: uploadResume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadResumeRequest'
      responses:
        '201':
          description: Resume uploaded and made active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedFile'
  /api/v2/resumes/{resumeId}:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    get:
      tags: [Resumes]
      summary: Get an active resume
      operationId: getResume
      responses:
        '200':
          description: User-owned or administrator-visible active resume metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    delete:
      tags: [Resumes]
      summary: Soft-delete an active resume
      operationId: softDeleteResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteResumeRequest'
      responses:
        '200':
          description: Resume soft-deleted and removed from active views.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
  /api/v2/recovery/resumes:
    get:
      tags: [Recovery]
      summary: List the current user's recoverable resumes
      operationId: listUserRecoveryResumes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Own user-soft-deleted or user-cache-archived resumes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /api/v2/recovery/resumes/{resumeId}/restore:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    post:
      tags: [Recovery]
      summary: Restore an eligible user resume
      operationId: restoreUserResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreResumeRequest'
      responses:
        '200':
          description: Resume restored to active visibility.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/RestoreConflict'
  /api/v2/admin/recovery/resumes:
    get:
      tags: [Administrator Recovery]
      summary: List recoverable resumes across owners
      operationId: listAdminRecoveryResumes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: ownerId
          in: query
          required: false
          description: Optional owner filter for an administrator.
          schema:
            type: string
            pattern: '^user[0-9]{3,}$'
      responses:
        '200':
          description: Recoverable resumes with owner context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /api/v2/admin/recovery/resumes/{resumeId}:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    delete:
      tags: [Administrator Recovery]
      summary: Soft-delete a resume as an administrator
      operationId: adminSoftDeleteResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteResumeRequest'
      responses:
        '200':
          description: Resume hidden by administrator soft deletion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
  /api/v2/admin/recovery/resumes/{resumeId}/restore:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    post:
      tags: [Administrator Recovery]
      summary: Restore a resume as an administrator
      operationId: adminRestoreResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreResumeRequest'
      responses:
        '200':
          description: Resume restored to active visibility by an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/RestoreConflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResumeId:
      name: resumeId
      in: path
      required: true
      description: Resume identifier.
      schema:
        type: string
        pattern: '^resume[0-9]{3,}$'
...
    Page:
      name: page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Number of records per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication is required or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: The authenticated actor lacks the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotFound:
      description: The resource does not exist or is not visible to this actor. Unknown and foreign IDs use this same envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundError'
          example:
            code: RESOURCE_NOT_FOUND
            message: Resource not found.
            correlationId: 00000000-0000-4000-8000-000000000001
            retryable: false
...
    VersionConflict:
      description: The optimistic-lock version is stale.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    RestoreConflict:
      description: The restore conflicts with a stale version, an existing effective resume title, or a resume that has no effective matching version. A duplicate title uses detailCode DUPLICATE_RESUME_TITLE and field title; a resume without any effective matching version uses code RESUME_NOT_EFFECTIVE.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    UnsupportedFile:
      description: The uploaded file type is unsupported in this slice.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    PayloadTooLarge:
      description: The uploaded file exceeds the configured size limit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, retryable]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - DUPLICATE_RESOURCE
            - INVALID_CONFIRMATION
            - VERSION_CONFLICT
            - RESOURCE_NOT_FOUND
            - RESUME_ARCHIVED
            - RESUME_NOT_EFFECTIVE
            - RESUME_SOFT_DELETED
            - TASK_GONE
            - STALE_ATTEMPT
            - IDEMPOTENCY_CONFLICT
            - MODEL_UNAVAILABLE
            - PYTHON_SERVICE_UNAVAILABLE
            - PYTHON_SERVICE_AUTHENTICATION_FAILED
            - MODEL_OUTPUT_INVALID
            - MODEL_ENDPOINT_REJECTED
            - CALLBACK_DELIVERY_FAILED
            - UNSUPPORTED_FILE
            - PAYLOAD_TOO_LARGE
            - TASK_NOT_READY
          example: RESUME_ARCHIVED
        message:
          type: string
          minLength: 1
          example: The resume is archived and must be restored first.
        correlationId:
          type: string
          format: uuid
        retryable:
          type: boolean
        detailCode:
          type: [string, 'null']
          enum: [DUPLICATE_RESUME_TITLE, null]
        details:
          type: array
          description: Safe validation details without protected content.
          items:
            $ref: '#/components/schemas/ValidationDetail'
      allOf:
        - if:
            required: [detailCode]
            properties:
              detailCode: { const: DUPLICATE_RESUME_TITLE }
          then:
            properties:
              code: { const: DUPLICATE_RESOURCE }
    ValidationDetail:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
          minLength: 1
        reason:
          type: string
          minLength: 1
    ResourceNotFoundError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          required: [code]
          properties:
            code:
              const: RESOURCE_NOT_FOUND
...
    VisibilityState:
      type: string
      enum: [ACTIVE, USER_SOFT_DELETED, ADMIN_SOFT_DELETED, USER_CACHE_ARCHIVED, ADMIN_CACHE_ARCHIVED]
...
    SourceType:
      type: string
      enum: [TXT, DOCX]
...
    UploadResumeRequest:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: UTF-8 TXT or DOCX document. PDF is unsupported in v1.
        title:
          type: string
          minLength: 1
          maxLength: 200
    Resume:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, sourceType, status, visibilityState, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: '^resume[0-9]{3,}$'
        ownerId:
          type: string
          pattern: '^user[0-9]{3,}$'
        title:
          type: string
        sourceType:
          $ref: '#/components/schemas/SourceType'
        status:
          type: integer
          enum: [0, 1]
          description: 0 means not soft-deleted; 1 means soft deletion completed.
        visibilityState:
          $ref: '#/components/schemas/VisibilityState'
        version:
          type: integer
          minimum: 0
        visibleUntil:
          type: [string, 'null']
          format: date-time
        softDeletedAt:
          type: [string, 'null']
          format: date-time
        archivedAt:
          type: [string, 'null']
          format: date-time
        restoredAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResumePage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Resume'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    DeleteResumeRequest:
      type: object
      additionalProperties: false
      required: [confirmationText, expectedVersion]
      properties:
        confirmationText:
          type: string
          const: '确认删除简历'
        expectedVersion:
          type: integer
          minimum: 0
    RestoreResumeRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
...